requirements.txt
discord.py
asyncio
aiohttp
zabbix_utils
